fix: modal example
